style(deck): move to deck package
