+++ src/main/kotlin/blackjack/domain/deck/Card.kt
@@ -1,4 +1,4 @@
-package blackjack.domain
+package blackjack.domain.deck
 
 data class Card(val suits: Suits, val denomination: Denomination) {
     override fun toString(): String {
